% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{board_url_manifest}
\alias{board_url_manifest}
\title{Create a URL board with a manifest}
\usage{
board_url_manifest(url, cache = NULL, use_cache_on_failure = is_interactive())
}
\arguments{
\item{url}{root URL of the board}

\item{cache}{Cache path. Every board requires a local cache to avoid
downloading files multiple times. The default stores in a standard
cache location for your operating system, but you can override if needed.}

\item{use_cache_on_failure}{If the pin fails to download, is it ok to
use the last cached version? Defaults to \code{is_interactive()} so you'll
be robust to poor internet connectivity when exploring interactively,
but you'll get clear errors when the code is deployed.}
}
\description{
Create a URL board with a manifest
}
\examples{
board <- board_url_manifest(url_demo_manifest())
pins::pin_list(board)
pins::pin_meta(board, "mtcars-csv")

}
