% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest_latest}
\alias{manifest_latest}
\title{Get the manifest of the latest pins at a URL}
\usage{
manifest_latest(url)
}
\arguments{
\item{url}{root URL of the board}
}
\value{
\code{character}, named vector:
\itemize{
\item names are names of pins
\item values
}
}
\description{
Reads a manifest created using \code{\link[=pin_write_manifest]{pin_write_manifest()}} to create a named
character vector that can be used to create a \code{\link[pins:board_url]{pins::board_url()}}.
}
\examples{
manifest_latest(url_demo_manifest())
}
