% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_write_manifest.R
\name{pin_write_manifest}
\alias{pin_write_manifest}
\title{Write a manifest file}
\usage{
pin_write_manifest(board)
}
\arguments{
\item{board}{A pin board, currently only \code{board_folder()} is supported.}
}
\value{
\code{board}, invisibly.
}
\description{
This can be used if you have a \code{pins::board_folder()} that you want to
serve as a web-site such that others can consume using \code{pins::board_url()}.
}
\details{
This function is called for the side-effect of writing a manifest file,
\verb{_pins.yaml}.
}
