% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinsManifest-package.R
\docType{package}
\name{pinsManifest-package}
\alias{pinsManifest}
\alias{pinsManifest-package}
\title{pinsManifest: Manage Manifests for URL Boards in the Pins Package}
\description{
The pins package supports URL boards, but it is difficult to manage the paths to the pins themselves. This package proposes a manifest file, which would be created as a part of a folder board, then used with a URL board when deployed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ijlyttle/pinsManifest}
  \item \url{https://ijlyttle.github.io/pinsManifest/}
  \item Report bugs at \url{https://github.com/ijlyttle/pinsManifest/issues}
}

}
\author{
\strong{Maintainer}: Ian Lyttle \email{ijlyttle@me.com} (\href{https://orcid.org/0000-0001-9962-4849}{ORCID})

Other contributors:
\itemize{
  \item Lionel Henry \email{lionel@rstudio.com} (compat-purrr) [contributor]
  \item Hadley Wickham \email{hadley@rstudio.com} (compat-purrr) [contributor]
}

}
\keyword{internal}
