% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{url_demo_manifest}
\alias{url_demo_manifest}
\title{Demonstration URL}
\usage{
url_demo_manifest()
}
\value{
\code{character} URL for demonstration board
}
\description{
Demonstration URL
}
\examples{
url_demo_manifest()
}
